requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers==0.35.2
transformers==4.47.0
accelerate>=0.25.0
safetensors>=0.4.0
omegaconf>=2.3.0
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0
huggingface-hub>=0.20.0
peft>=0.17.0

# Web API
flask>=3.0.0
flask-cors>=4.0.0
flask-login>=0.6.3
flask-talisman>=1.1.0
flask-sqlalchemy>=3.0.0
authlib>=1.3.0
gunicorn>=21.2.0
pyyaml>=6.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# Web Scraping
playwright>=1.40.0
httpx>=0.25.0
beautifulsoup4>=4.12.0
lxml>=4.9.3
imagehash>=4.3.1
aiofiles>=23.2.1
pydantic>=2.5.0
pydantic-settings>=2.1.0

# AI Labeling & Captioning
# Uses Claude CLI via Docker (no direct API dependency)

# Task Queue
celery>=5.3.0
redis>=5.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
jsonschema>=4.21.1

# Linting
black>=25.0.0
flake8>=7.0.0
isort>=7.0.0
